+++ src/Modules/RenderList.js
@@ -3,9 +3,9 @@ import {
   storage,
   changeColor,
   editList,
-  deleteElement,
 } from './addRemove.js';
 import addList from './addList.js';
+import deleteElement from './deleteItem.js';
 
 function renderList() {
   const listItems = storage();
